oracle/requirements.txt
compress_json==1.0.10
langchain==0.1.0
langchain_community==0.0.13
langchain_openai==0.0.2.post1
tqdm==4.65.0
python_dotenv==1.0.0
unstructured==0.12.3
faiss-cpu==1.7.4
